requirements.txt
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
python-dotenv>=1.0